ci: create another deploy workflow to publish _site to S3 in production
deploy-s3-prod will publish _site only when something is pushed to master
branch.

deploy-s3 will pushed _site only when a PR has been created and the target
branch is `dev`. it does not when you push a branch. as such, you cannot see
your changes in your branch before creating a PR.

the name of the bucket for production is `mywebtest-prod`.

the bucket for the production environment must exist.

fixes #10